Refactor rgb_light
