+++ MLight/MLight.slcp
@@ -198,8 +198,8 @@ source:
   - path: app.c
   - path: mods/device-nwk-join-control.h
   - path: mods/device-nwk-join-control.c
-  - path: hw/rgb_light.h
-  - path: hw/rgb_light.c
+  - path: light/hw_light.h
+  - path: light/hw_light.c
 
 config_file:
   - path: config/zcl/zcl_config.zap
